Fix reading of parameter defaults
There is a bug in dynamic_reconfigure: The config.groups.* fields don't
get initialized with the proper defaults, but the config.* fields are.
This commit makes sure that the defaults are properly read from the
parameter server (or from cfg/CameraParameters.cfg if not set on the
parameter server), so we don't need to call the dynamic reconfigure
callback explicitly after starting the node.